% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_clr_cpp}
\alias{apply_clr_cpp}
\title{Apply CLR (Context Likelihood Ratio) transformation}
\usage{
apply_clr_cpp(mi_matrix, n_cores = 1L)
}
\arguments{
\item{mi_matrix}{Symmetric mutual information matrix}

\item{n_cores}{Number of OpenMP threads (currently unused, for API consistency)}
}
\value{
CLR-transformed matrix
}
\description{
Normalizes MI matrix using z-scores to reduce hub gene bias.
}
\details{
CLR transformation (Faith et al. 2007):
\preformatted{
Z_row(i,j) = (MI(i,j) - mean(row_i)) / sd(row_i)
Z_col(i,j) = (MI(i,j) - mean(col_j)) / sd(col_j)
CLR(i,j) = sqrt(max(0, Z_row)^2 + max(0, Z_col)^2)
}
}
