% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mi_clr_cpp}
\alias{compute_mi_clr_cpp}
\title{Compute MI+CLR similarity matrix (full matrix output)}
\usage{
compute_mi_clr_cpp(expr, n_bins, n_cores = 1L)
}
\arguments{
\item{expr}{Expression matrix (genes x samples)}

\item{n_bins}{Number of bins for discretization}

\item{n_cores}{Number of OpenMP threads}
}
\value{
CLR-transformed similarity matrix
}
\description{
Full pipeline: discretize -> MI -> CLR, returning full matrix.
}
