% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mi_matrix_cpp}
\alias{compute_mi_matrix_cpp}
\title{Compute mutual information matrix}
\usage{
compute_mi_matrix_cpp(expr_disc, n_bins, n_cores = 1L)
}
\arguments{
\item{expr_disc}{Discretized expression matrix (genes x samples, integer)}

\item{n_bins}{Number of bins used in discretization}

\item{n_cores}{Number of OpenMP threads}
}
\value{
Symmetric MI matrix (genes x genes)
}
\description{
Computes pairwise MI for all gene pairs from discretized expression.
}
\details{
This function computes MI(i,j) = H(i) + H(j) - H(i,j) for all gene pairs.
The diagonal contains self-information (entropy).
Parallelized via OpenMP for efficiency.
}
