% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretize_equalfreq_cpp}
\alias{discretize_equalfreq_cpp}
\title{Discretize a vector into equal-frequency bins}
\usage{
discretize_equalfreq_cpp(x, n_bins)
}
\arguments{
\item{x}{Numeric vector to discretize}

\item{n_bins}{Number of bins}
}
\value{
Integer vector of bin assignments (0-indexed)
}
\description{
Discretize a vector into equal-frequency bins
}
