% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretize_matrix_cpp}
\alias{discretize_matrix_cpp}
\title{Discretize expression matrix (all genes)}
\usage{
discretize_matrix_cpp(expr, n_bins, n_cores = 1L)
}
\arguments{
\item{expr}{Expression matrix (genes x samples)}

\item{n_bins}{Number of bins}

\item{n_cores}{Number of OpenMP threads (default 1)}
}
\value{
Integer matrix of bin assignments (genes x samples)
}
\description{
Discretize expression matrix (all genes)
}
