% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{extractColumn}
\alias{extractColumn}
\alias{extractColumn,TriSimilarity-method}
\title{Extract a column from TriSimilarity}
\usage{
extractColumn(x, gene)

\S4method{extractColumn}{TriSimilarity}(x, gene)
}
\arguments{
\item{x}{TriSimilarity object}

\item{gene}{Gene name or index to extract column for}
}
\value{
Named numeric vector
}
\description{
Efficiently extracts a full column (co-expression vector) from a
triangular similarity matrix.
}
