% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{extractColumns}
\alias{extractColumns}
\alias{extractColumns,TriSimilarity-method}
\title{Extract multiple columns from TriSimilarity}
\usage{
extractColumns(x, genes)

\S4method{extractColumns}{TriSimilarity}(x, genes)
}
\arguments{
\item{x}{TriSimilarity object}

\item{genes}{Gene names or indices to extract}
}
\value{
Numeric matrix with rows = all genes, columns = requested genes
}
\description{
Extract multiple columns from TriSimilarity
}
