% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{extractRows}
\alias{extractRows}
\alias{extractRows,TriSimilarity-method}
\title{Extract rows from TriSimilarity}
\usage{
extractRows(x, genes)

\S4method{extractRows}{TriSimilarity}(x, genes)
}
\arguments{
\item{x}{TriSimilarity object}

\item{genes}{Gene names or indices to extract}
}
\value{
Numeric matrix with rows = requested genes, columns = all genes
}
\description{
Extract rows from TriSimilarity
}
