% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{is.TriSimilarity}
\alias{is.TriSimilarity}
\title{Check if object is TriSimilarity}
\usage{
is.TriSimilarity(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is TriSimilarity
}
