% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{loadTriSimilarity}
\alias{loadTriSimilarity}
\title{Load TriSimilarity from file}
\usage{
loadTriSimilarity(file)
}
\arguments{
\item{file}{File path}
}
\value{
TriSimilarity object
}
\description{
Load TriSimilarity from file
}
