% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ccs_vs_ors}
\alias{plot_ccs_vs_ors}
\title{Plot CCS vs logORS}
\usage{
plot_ccs_vs_ors(ors_results, title = "CCS vs logORS")
}
\arguments{
\item{ors_results}{Data frame from \code{\link[=calculate_ors]{calculate_ors()}} with \code{CCS} and
\code{logORS} columns.}

\item{title}{Character. Plot title. Default "CCS vs logORS".}
}
\value{
A ggplot2 object.
}
\description{
Scatter plot of CCS against logORS with conservation zones colored.
}
\examples{
\dontrun{
ors_results <- calculate_ors(ccs_results)
plot_ccs_vs_ors(ors_results)
}

}
