% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ors_by_type}
\alias{plot_ors_by_type}
\title{Plot logORS by ortholog type}
\usage{
plot_ors_by_type(ors_results, title = "Conservation by ortholog type")
}
\arguments{
\item{ors_results}{Data frame from \code{\link[=calculate_ors]{calculate_ors()}} with \code{logORS} and
\code{type} columns.}

\item{title}{Character. Plot title. Default "Conservation by ortholog type".}
}
\value{
A ggplot2 object.
}
\description{
Boxplot comparing logORS distributions across ortholog types (1:1, 1:N,
N:1, N:M). Requires a \code{type} column in the input data.
}
\examples{
\dontrun{
ors_results <- calculate_ors(ccs_results)
plot_ors_by_type(ors_results)
}

}
