% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{saveTriSimilarity}
\alias{saveTriSimilarity}
\title{Save TriSimilarity to file}
\usage{
saveTriSimilarity(x, file)
}
\arguments{
\item{x}{TriSimilarity object}

\item{file}{File path (should end in .rds)}
}
\value{
Invisibly returns file path
}
\description{
Save TriSimilarity to file
}
