% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{[,TriSimilarity,ANY,ANY-method}
\alias{[,TriSimilarity,ANY,ANY-method}
\alias{[,TriSimilarity,ANY,missing-method}
\alias{[,TriSimilarity,missing,ANY-method}
\title{Subset a TriSimilarity object}
\usage{
\S4method{[}{TriSimilarity,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{TriSimilarity,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{TriSimilarity,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{TriSimilarity object}

\item{i}{Row indices or gene names}

\item{j}{Column indices or gene names}

\item{...}{Ignored}

\item{drop}{Logical. If TRUE and result is single value, return scalar.}
}
\value{
Numeric value(s) or matrix subset.
}
\description{
Subset a TriSimilarity object
}
