% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\docType{class}
\name{TriSimilarity-class}
\alias{TriSimilarity-class}
\alias{as.matrix,TriSimilarity-method}
\alias{dim,TriSimilarity-method}
\alias{dimnames,TriSimilarity-method}
\alias{show,TriSimilarity-method}
\title{TriSimilarity Class}
\usage{
\S4method{as.matrix}{TriSimilarity}(x, ...)

\S4method{dim}{TriSimilarity}(x)

\S4method{dimnames}{TriSimilarity}(x)

\S4method{show}{TriSimilarity}(object)
}
\arguments{
\item{x}{TriSimilarity object}

\item{...}{Ignored (for S4 method consistency)}

\item{object}{TriSimilarity object}
}
\description{
S4 class for efficient storage of symmetric similarity matrices.
Stores only the upper triangle (excluding diagonal), reducing memory
by approximately 50\%.
}
\details{
The upper triangle is stored in column-major order, matching R's
\code{upper.tri()} ordering. For a matrix with n genes, the data
slot contains n*(n-1)/2 elements.

For position (i,j) with i < j, the index in the data vector is:
\code{(j-1)*(j-2)/2 + i}
}
\section{Functions}{
\itemize{
\item \code{as.matrix(TriSimilarity)}: Convert to full matrix

\item \code{dim(TriSimilarity)}: Get dimensions

\item \code{dimnames(TriSimilarity)}: Get dimension names

\item \code{show(TriSimilarity)}: Show method

}}
\section{Slots}{

\describe{
\item{\code{data}}{Numeric vector of upper triangle values in column-major order.}

\item{\code{genes}}{Character vector of gene names.}

\item{\code{n}}{Integer. Number of genes.}

\item{\code{diag_value}}{Numeric. Value for diagonal elements (typically 1.0).}
}}

\seealso{
\code{\link{TriSimilarity}} for the constructor function.
}
