% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriSimilarity-class.R
\name{TriSimilarity}
\alias{TriSimilarity}
\title{Create a TriSimilarity object}
\usage{
TriSimilarity(data, genes, diag_value = 1)
}
\arguments{
\item{data}{Numeric vector of upper triangle values (column-major order,
excluding diagonal). Length should be n*(n-1)/2 where n is number of genes.}

\item{genes}{Character vector of gene names.}

\item{diag_value}{Numeric. Value for diagonal elements (default 1.0).}
}
\value{
A TriSimilarity object.
}
\description{
Constructor for triangular similarity matrices.
}
\examples{
\dontrun{
# Create from components
tri <- TriSimilarity(
  data = c(0.5, 0.3, 0.7),  # Upper triangle of 3x3
  genes = c("A", "B", "C"),
  diag_value = 1.0
)
}

}
