% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_orthologs}
\alias{validate_orthologs}
\title{Validate ortholog table format}
\usage{
validate_orthologs(orthologs)
}
\arguments{
\item{orthologs}{Ortholog data frame to validate}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate ortholog table format
}
\keyword{internal}
